+++ sdk/computerecommender/azure-mgmt-computerecommender/generated_samples/generate_spot_placement_scores.py
@@ -42,6 +42,6 @@ def main():
     print(response)
 
 
-# x-ms-original-file: 2025-06-05/GenerateSpotPlacementScores.json
+# x-ms-original-file: 2026-05-05-preview/GenerateSpotPlacementScores.json
 if __name__ == "__main__":
     main()