+++ sdk/network/azure-mgmt-network/generated_tests/test_network_management_routes_operations_async.py
@@ -26,7 +26,7 @@ class TestNetworkManagementRoutesOperationsAsync(AzureMgmtRecordedTestCase):
             resource_group_name=resource_group.name,
             route_table_name="str",
             route_name="str",
-            api_version="2025-07-01",
+            api_version="2025-09-01",
         )
 
         # please add some check logic here by yourself
@@ -54,7 +54,7 @@ class TestNetworkManagementRoutesOperationsAsync(AzureMgmtRecordedTestCase):
                     },
                     "type": "str",
                 },
-                api_version="2025-07-01",
+                api_version="2025-09-01",
             )
         ).result()  # call '.result()' to poll until service return final result
 
@@ -69,7 +69,7 @@ class TestNetworkManagementRoutesOperationsAsync(AzureMgmtRecordedTestCase):
                 resource_group_name=resource_group.name,
                 route_table_name="str",
                 route_name="str",
-                api_version="2025-07-01",
+                api_version="2025-09-01",
             )
         ).result()  # call '.result()' to poll until service return final result
 
@@ -82,7 +82,7 @@ class TestNetworkManagementRoutesOperationsAsync(AzureMgmtRecordedTestCase):
         response = self.client.routes.list(
             resource_group_name=resource_group.name,
             route_table_name="str",
-            api_version="2025-07-01",
+            api_version="2025-09-01",
         )
         result = [r async for r in response]
         # please add some check logic here by yourself