+++ sdk/sql/azure-mgmt-sql/generated_samples/managed_database_sensitivity_labels_list_by_database.py
@@ -39,6 +39,6 @@ def main():
         print(item)
 
 
-# x-ms-original-file: 2025-01-01/ManagedDatabaseSensitivityLabelsListByDatabase.json
+# x-ms-original-file: 2025-08-01-preview/ManagedDatabaseSensitivityLabelsListByDatabase.json
 if __name__ == "__main__":
     main()