+++ sdk/sql/azure-mgmt-sql/generated_samples/start_stop_managed_instance_schedule_create_or_update_max.py
@@ -48,6 +48,6 @@ def main():
     print(response)
 
 
-# x-ms-original-file: 2025-01-01/StartStopManagedInstanceScheduleCreateOrUpdateMax.json
+# x-ms-original-file: 2025-08-01-preview/StartStopManagedInstanceScheduleCreateOrUpdateMax.json
 if __name__ == "__main__":
     main()